# conduit-api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Conduit API",
    "description": "REST API for the Conduit data pipeline orchestrator. Conduit is a Rust-native orchestration engine for DAGs (directed acyclic graphs) that compiles and schedules data pipelines with support for complex dependencies, dynamic scheduling, and schema contracts.",
    "version": "0.1.0",
    "contact": {
      "name": "Conduit Community"
    }
  },
  "servers": [
    {
      "url": "http://localhost:9000/api/v1",
      "description": "Local development server"
    },
    {
      "url": "https://api.conduit.dev/api/v1",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/lineage/graph": {
      "post": {
        "tags": ["Lineage"],
        "summary": "Get lineage graph",
        "description": "Build complete lineage graph for visualization.",
        "operationId": "lineageGraph",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lineage graph",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nodes": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "edges": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/contracts/{dagId}": {
      "get": {
        "tags": ["Contracts"],
        "summary": "Get DAG contracts",
        "operationId": "dagContracts",
        "parameters": [
          {
            "name": "dagId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "DAG contracts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contracts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ContractSummary"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/contracts/{dagId}/{taskId}": {
      "get": {
        "tags": ["Contracts"],
        "summary": "Get task contracts",
        "operationId": "taskContracts",
        "parameters": [
          {
            "name": "dagId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task contracts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contracts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ContractSummary"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ApiKey",
        "description": "API key authentication. Use format: Authorization: Bearer <api_key>"
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ContractSummary": {
        "type": "object",
        "properties": {
          "dagId": {
            "type": "string"
          },
          "taskId": {
            "type": "string"
          },
          "checkCount": {
            "type": "integer"
          },
          "checks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "checkType": {
                  "type": "string"
                },
                "severity": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  }
}